refactor: index.js를 이용해서 import문 간소화, RoundCount -> Race로 클래스 이름 변환
