+++ src/models/index.js
@@ -1,0 +1,2 @@
+export { default as Car } from './Car';
+export { default as Race } from './Race';
