+++ src/utils/index.js
@@ -1,0 +1,4 @@
+export { default as Console } from './Console';
+export { default as Validator } from './Validator';
+export { default as gameUtils } from './gameUtils';
+export { default as repeatFunctionUntilIsValid } from './repeatFunctionUntilIsValid';
