+++ src/views/InputView.js
@@ -1,4 +1,4 @@
-import Console from '../utils/Console';
+import { Console } from '../utils';
 
 const InputView = {
   async readCarNames() {
