+++ src/views/index.js
@@ -1,0 +1,2 @@
+export { default as InputView } from './InputView';
+export { default as OutputView } from './OutputView';
